refactor: moves CODEOWNERS related sources to their own folder
