--- src/__fixtures__/corpus/empty-teams/000-empty.yml
@@ -1,1 +1,0 @@
-[{ type: "empty", line: 1, raw: "" }]
