--- src/__fixtures__/corpus/empty-teams/001-comment.txt
@@ -1,1 +1,0 @@
-# a comment
