--- src/__fixtures__/corpus/empty-teams/001-comment.yml
@@ -1,1 +1,0 @@
-[{ type: "comment", line: 1, raw: "# a comment" }]
