--- src/__fixtures__/corpus/empty-teams/002-ignorable-comment.txt
@@ -1,1 +1,0 @@
-#! an ignorable comment
