--- src/__fixtures__/corpus/empty-teams/002-ignorable-comment.yml
@@ -1,1 +1,0 @@
-[{ type: "ignorable-comment", line: 1, raw: "#! an ignorable comment" }]
