--- src/__fixtures__/corpus/empty-teams/003-unparsable-line.txt
@@ -1,1 +1,0 @@
-this-is-not-a-valid-rule
