--- src/__fixtures__/corpus/empty-teams/003-unparsable-line.yml
@@ -1,1 +1,0 @@
-[{ type: "unknown", line: 1, raw: "this-is-not-a-valid-rule" }]
