--- src/__fixtures__/corpus/empty-teams/010-rule-other-user.txt
@@ -1,1 +1,0 @@
-* @everyone
