--- src/__fixtures__/corpus/empty-teams/011-rule-e-mail-user.txt
@@ -1,1 +1,0 @@
-* jan@example.com
