--- src/__fixtures__/corpus/empty-teams/012-rule-invalid-user.txt
@@ -1,1 +1,0 @@
-* username-does-not-start-with-an-at
