--- src/__fixtures__/corpus/empty-teams/013-rule-with-inline-comment.txt
@@ -1,1 +1,0 @@
-* @everyone   # inline comment
