--- src/__fixtures__/corpus/empty-teams/014-rule-with-non-space-sepped-inline-comment.txt
@@ -1,1 +1,0 @@
-* @everyone# inline comment
