--- src/__fixtures__/corpus/empty-teams/014-rule-with-non-space-sepped-inline-comment.yml
@@ -1,19 +1,0 @@
-[
-  {
-    type: "rule",
-    line: 1,
-    filesPattern: "*",
-    spaces: " ",
-    users:
-      [
-        {
-          type: "other-user-or-team",
-          raw: "@everyone",
-          bareName: "everyone",
-          "userNumberWithinLine": 1,
-        },
-      ],
-    inlineComment: " inline comment",
-    raw: "* @everyone# inline comment",
-  },
-]
