--- src/__fixtures__/corpus/teams/020-rule-virtual-team.txt
@@ -1,1 +1,0 @@
-* @baarden
