--- src/__fixtures__/corpus/teams/020-rule-virtual-team.yml
@@ -1,19 +1,0 @@
-[
-  {
-    type: "rule",
-    line: 1,
-    filesPattern: "*",
-    spaces: " ",
-    users:
-      [
-        {
-          type: "virtual-team-name",
-          raw: "@baarden",
-          bareName: "baarden",
-          userNumberWithinLine: 1,
-        },
-      ],
-    inlineComment: "",
-    raw: "* @baarden",
-  },
-]
