--- src/__fixtures__/corpus/teams/021-rule-multiple-teams.txt
@@ -1,1 +1,0 @@
-* @baarden @leren-lezen @real-team jet@example.com
