--- src/__fixtures__/corpus/teams/021-rule-multiple-teams.yml
@@ -1,37 +1,0 @@
-[
-  {
-    type: "rule",
-    line: 1,
-    filesPattern: "*",
-    spaces: " ",
-    users:
-      [
-        {
-          type: "virtual-team-name",
-          raw: "@baarden",
-          bareName: "baarden",
-          userNumberWithinLine: 1,
-        },
-        {
-          type: "virtual-team-name",
-          raw: "@leren-lezen",
-          bareName: "leren-lezen",
-          userNumberWithinLine: 2,
-        },
-        {
-          type: "other-user-or-team",
-          raw: "@real-team",
-          bareName: "real-team",
-          userNumberWithinLine: 3,
-        },
-        {
-          type: "e-mail-address",
-          raw: "jet@example.com",
-          bareName: "jet@example.com",
-          userNumberWithinLine: 4,
-        },
-      ],
-    inlineComment: "",
-    raw: "* @baarden @leren-lezen @real-team jet@example.com",
-  },
-]
