+++ src/codeowners/__mocks__/virtual-teams.yml
@@ -1,0 +1,39 @@
+# yaml-language-server: $schema=../virtual-teams.schema.json
+ch/after-sales:
+  - john-doe-ch
+  - pete-peterson-ch
+  - william-the-fourth-ch
+  - daisy-duck
+  - donald-duck
+ch/sales:
+  - gregory-gregson-ch
+  - jane-doe-ch
+  - abraham-ableton-ch
+  - boris-bubbleblower-ch
+  - charlotte-charleston-ch
+  - dagny-taggert-ch
+  - karl-marx-ch
+ch/pre-sales:
+  - jean-claude-ch
+  - valerie-valerton-ch
+  - averel-dalton-ch
+  - john-galt-ch
+ch/ux:
+  - davy-davidson-ch
+  - john-johnson-ch
+  - joe-dalton-ch
+  - korneel@example.com
+  - koos-koets
+  - charlotte-de-bourbon-ch
+ch/transversal:
+  - mary-the-merry-ch
+  - luke-the-lucky-ch
+  - naomi-the-namegiver-ch
+  - benjamin-franklin
+  - koos-koets
+  - abraham-lincoln
+ch/mannen-met-baarden:
+  - jan@example.com
+  - pier@example.com
+  - tjorus@example.com
+  - korneel@example.com
