+++ src/codeowners/generate.ts
@@ -5,7 +5,7 @@ import type {
   IVirtualCodeOwnerLine,
   IVirtualCodeOwnersCST,
 } from "types/types.js";
-import { isEmailIshUsername } from "./utensils.js";
+import { isEmailIshUsername } from "../utensils.js";
 
 const DEFAULT_WARNING =
   `#${EOL}` +
